requirements/base.txt
asgiref==3.7.2
Django==4.2.3
django-environ==0.9.0
django-filter==23.2
djangorestframework==3.14.0
psycopg2-binary==2.9.3  # for Ubuntu 23.04 2.9.7
python-dotenv==0.21.0
sqlparse==0.4.4
typing_extensions==4.7.1
